requirements.txt
streamlit
bs4
googlesearch-python